Strip whitespace from attributes when importing transient registration
If the attributes we're importing from a registration have excess whitespace at the start or end, clean them up before importing them.

Decided to move the whitespace stripping methods to a concern as they're being shared by the form objects and the TransientRegistration model.

Also removed unnecessary helper file.
